FPLib/requirements.txt
pytz==2021.1
Flask==1.1.4
tqdm==4.61.0
psutil==5.8.0
paramiko==2.7.2
pyzbar==0.1.8
gevent==21.1.2
urllib3==1.26.4
beautifulsoup4==4.9.3
Pillow==8.3.1
